Mostly Refactor Dynamic
Isolates the finisher and update_latest functions with their distinct dynamic website handling purpose

Introduces a dest_VAR methodology for QtWebEngine to access otherwise impossible data values

TODO: Verify smooth runtime

TODO: Alter checker logic to pass more arguments and accept updated output
