+++ .gitignore
@@ -11,3 +11,4 @@ __pycache__
 walkthrough
 user.txt
 access/credentials.json
+tester*
