+++ saved/list.txt
@@ -1,4 +1,4 @@
-Solo Leveling (1) [System Leveling]|https://www.asurascans.com/manga/4569947261-solo-leveling/|asura|
+Solo Leveling (1) [System Leveling]|https://www.asurascans.com/manga/9260952888-solo-leveling/|asura|
 Tales of Demons and Gods (2) [Cultivator Reincarnation]|https://manhuaplus.com/manga/tales-of-demons-and-gods/|WP|
 The Great Mage Returns After 4000 Years (2) [Cultivator Reincarnation]|https://www.mangaread.org/manga/the-great-mage-returns-after-4000-years/|MR|
 Second Life Ranker (1) [Cultivator Reincarnation]|https://www.mangaread.org/manga/ranker-who-lives-a-second-time/|MR|
@@ -11,4 +11,4 @@ Demon Magic Emperor (3) [Cultivator Reincarnation]|https://manhuaplus.com/manga/
 Leveling Up, by Only Eating! (1) [Virtual Reality Game]|https://www.mangaread.org/manga/leveling-up-by-only-eating/|WP|
 Yuan Zun (1) [Underdog Cultivator]|https://manhuaplus.com/manga/yuan-zun/|WP|
 Spirit Farmer (2) [System Leveling]|https://mangaeffect.com/manga/spirit-farmer/|MR|
-The Greatest Estate Developer (1) [System Leveling]|https://www.asurascans.com/manga/4569947261-the-greatest-estate-developer/|asura|
+The Greatest Estate Developer (1) [System Leveling]|https://www.asurascans.com/manga/9260952888-the-greatest-estate-developer/|asura|
